Implements basic placeholder obstacle avoidance for enemies
* Enriches Penguin lives through the magic of dance
* Adds a function in MovingObject to raycast from current position
without attempting movement.
* Adds movement randomization when enemies are directly diagonal to the
player when seeking so that they don't always travel in the vertical
direction.
* Adds very basic obstacle avoidance by causing penguins to move in a
totally random direction when they would otherwise collide with an
obstacle or fellow penguin.

Makes enemies decide on a coin flip which direction to travel if they are directly diagonal to the player when seeking.

Adds a small comment

Creates idiotic pathing around obstacles

Removes debug prints to console